refactor: 2단계 1차 피드백 반영 - 로또 번호 일치 개수, 당첨 금액 LottoResult로 분리
