+++ src/main/java/lotto/Lotto.java
@@ -3,23 +3,25 @@ package lotto;
 import java.util.*;
 
 public class Lotto {
+    private static final int LOTTO_NUMBER_COUNT = 6;
     private static final Random RANDOM = new Random();
-    private static final List<Integer> MATCHING_COUNT_FOR_WINNING = Arrays.asList(3, 4, 5, 6);
 
     private List<Integer> lottoNumbers;
-    private int matchCount;
+    private LottoResult lottoResult;
 
     public Lotto() {
         lottoNumbers = new ArrayList<>();
         lottoNumbers = generateAutoLottoNumbers();
+        lottoResult = new LottoResult();
     }
 
     public Lotto(List<Integer> lottoNumbers) {
         this.lottoNumbers = lottoNumbers;
+        lottoResult = new LottoResult();
     }
 
     public int getMatchCount() {
-        return this.matchCount;
+        return this.lottoResult.getMatchCount();
     }
 
     public String toStringLottoNumbers() {
@@ -28,7 +30,7 @@ public class Lotto {
     }
 
     public List<Integer> generateAutoLottoNumbers() {
-        while (lottoNumbers.size() < 6) {
+        while (lottoNumbers.size() < LOTTO_NUMBER_COUNT) {
             generateLottoNumber();
         }
 
@@ -44,41 +46,15 @@ public class Lotto {
     }
 
     public int calculateMatchCount(List<Integer> winningNumbers) {
-        for (int i = 0; i < 6; i++) {
-            matchCount += checkMatchNumber(winningNumbers, lottoNumbers.get(i));
-        }
-        return matchCount;
+        return this.lottoResult.calculateMatchCount(winningNumbers, lottoNumbers);
     }
 
-    private int checkMatchNumber(List<Integer> winningNumbers, int number) {
-        if (winningNumbers.contains(number)) {
-            return 1;
-        }
-        return 0;
-    }
-
-    public int getAmountForWinningLotto() {
-        if (!this.isWinningLotto()) {
-            return 0;
-        }
-        return determineAmountByMatchCount(this.matchCount);
+    public int getAmount() {
+        return this.lottoResult.getAmountForWinningLotto();
     }
 
     public boolean isWinningLotto() {
-        return MATCHING_COUNT_FOR_WINNING.contains(this.matchCount);
-    }
-
-    public static int determineAmountByMatchCount(int matchCount) {
-        if (matchCount == 3) {
-            return 5000;
-        } else if (matchCount == 4) {
-            return 50000;
-        } else if (matchCount == 5) {
-            return 1500000;
-        } else if (matchCount == 6) {
-            return 2000000000;
-        }
-        return 0;
+        return this.lottoResult.isWinningLotto();
     }
 
     public static boolean validateLottoNumbers(List<Integer> numbers) { //todo private
